src/kserve_model/requirements.txt
onnxruntime
kserve
pillow
numpy